fix class, property, file name
