+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/TopActivity.kt
