+++ app/src/main/res/layout/repository_detail_fragment.xml
