+++ app/src/main/res/navigation/nav_graph.xml
@@ -9,7 +9,7 @@
         android:id="@+id/oneFragment"
         android:name="jp.co.yumemi.android.codeCheck.OneFragment"
         android:label="@string/app_name"
-        tools:layout="@layout/fragment_one">
+        tools:layout="@layout/search_repository_fragment">
         <action
             android:id="@+id/action_repositoriesFragment_to_repositoryFragment"
             app:destination="@id/twoFragment" />
@@ -17,12 +17,12 @@
 
     <fragment
         android:id="@+id/twoFragment"
-        android:name="jp.co.yumemi.android.codeCheck.TwoFragment"
+        android:name="jp.co.yumemi.android.codeCheck.RepositoryDetailFragment"
         android:label="@string/app_name"
-        tools:layout="@layout/fragment_two">
+        tools:layout="@layout/repository_detail_fragment">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.code_check.item" />
+            app:argType="jp.co.yumemi.android.codeCheck.Item" />
     </fragment>
 
 </navigation>
